+++ src/main/java/rhss_server/rhss_server/Interfaces/IUsuarioRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import rhss_server.rhss_server.Tables.UsuarioModel;
 
-public interface IUsuarioRepo extends JpaRepository<UsuarioModel, Byte> {
+public interface IUsuarioRepo extends JpaRepository<UsuarioModel, Long> {
 
     Optional<UsuarioModel> findByUsername(String username);
 }
